ctbriefs.net original code
This is the ctbriefs.net code after moving it to a new username on dreamhost.

At present, the site displays a blank white page.  I remember it had
some problems with something; maybe I killed it.  Anyway, I don't
expect to revive it, but I'm just going to throw it on github for fun.